feat: LGPD

+++ src/app/context/AuthenticationContext.tsx
@@ -16,6 +16,7 @@ export type Auth = {
 type LocalStorage = {
     isLogged: boolean;
     loggedUser: User | null;
+    allowCookies?: boolean;
 }
 
 export const AuthenticationContext = createContext<Auth | null>(null);
@@ -23,7 +24,9 @@ const LOCAL_STORAGE_KEY = "logged_user";
 
 export const AuthenticationProvider: React.FC<Props> = ({ children }) => {
     const [localStorageData, setLocalStorageData] = useState<LocalStorage>({
-        isLogged: false, loggedUser: {
+        allowCookies: false,
+        isLogged: false, 
+        loggedUser: {
             id: 0,
             name: "",
             last_name: "",
@@ -33,29 +36,33 @@ export const AuthenticationProvider: React.FC<Props> = ({ children }) => {
     });
 
     const loginManager = async () => {
-        const response = await api.check();
-        if (response !== undefined && response.message === "Authenticated") {
-            localStorage.setItem(LOCAL_STORAGE_KEY, JSON.stringify(response));
-            setLocalStorageData({
-                isLogged: true,
-                loggedUser: {
-                    id: response.logged_user.id,
-                    name: response.logged_user.name,
-                    last_name: response.logged_user.last_name,
-                    email: response.logged_user.email,
-                    telephone: response.logged_user.telephone,
-                }
-            });
-        } else {
-            setLocalStorageData({ isLogged: false, loggedUser: null });
+        if (localStorage.getItem("allowCookies")) {
+            const response = await api.check();
+            if (response !== undefined && response.message === "Authenticated") {
+                localStorage.setItem(LOCAL_STORAGE_KEY, JSON.stringify(response));
+                setLocalStorageData({
+                    isLogged: true,
+                    loggedUser: {
+                        id: response.logged_user.id,
+                        name: response.logged_user.name,
+                        last_name: response.logged_user.last_name,
+                        email: response.logged_user.email,
+                        telephone: response.logged_user.telephone,
+                    }
+                });
+            } else {
+                setLocalStorageData({ isLogged: false, loggedUser: null });
+            }
         }
     }
 
     const logout = async () => {
-        const response = await api.destroy();
-        if (response !== undefined && response.message === "The session was finished.") {
-            setLocalStorageData({ isLogged: false, loggedUser: null });
-            localStorage.removeItem(LOCAL_STORAGE_KEY);
+        if (localStorage.getItem("allowCookies")) {
+            const response = await api.destroy();
+            if (response !== undefined && response.message === "The session was finished.") {
+                setLocalStorageData({ isLogged: false, loggedUser: null });
+                localStorage.removeItem(LOCAL_STORAGE_KEY);
+            }
         }
     }
